cleanup: remove outdated docs, rename atomic_broadcast to aleph_bft
